refactor: move reserved keywords into the language generators
The Generator interface now extends Keywords, so each language's
generator is the single source of truth for its reserved keywords.
The IrEmitter companions delegate to their generator (Keywords by
XxxGenerator) and the Avro extension reads the generator sets, replacing
the standalone ReservedKeywords objects.
