+++ src/compiler/emitters/python/src/commonMain/kotlin/community/flock/wirespec/emitters/python/PythonIrEmitter.kt
@@ -39,7 +39,6 @@ import community.flock.wirespec.ir.core.raw
 import community.flock.wirespec.ir.core.transform
 import community.flock.wirespec.ir.core.transformChildren
 import community.flock.wirespec.ir.emit.IrEmitter
-import community.flock.wirespec.ir.generator.PythonKeywords
 import community.flock.wirespec.ir.emit.placeInModule
 import community.flock.wirespec.ir.emit.prependImports
 import community.flock.wirespec.ir.generator.PythonGenerator
@@ -309,6 +308,6 @@ open class PythonIrEmitter(
         import(wirespecPath, "_raise"),
     )
 
-    companion object : Keywords by PythonKeywords
+    companion object : Keywords by PythonGenerator
 
 }
