+++ src/compiler/ir/src/commonMain/kotlin/community/flock/wirespec/ir/generator/JavaGenerator.kt
@@ -60,6 +60,20 @@ import community.flock.wirespec.ir.core.fieldList
 import community.flock.wirespec.ir.core.Function as AstFunction
 
 object JavaGenerator : Generator {
+    override val reservedKeywords = setOf(
+        "abstract", "continue", "for", "new", "switch",
+        "assert", "default", "if", "package", "synchronized",
+        "boolean", "do", "goto", "private", "this",
+        "break", "double", "implements", "protected", "throw",
+        "byte", "else", "import", "public", "throws",
+        "case", "enum", "instanceof", "return", "transient",
+        "catch", "extends", "int", "short", "try",
+        "char", "final", "interface", "static", "void",
+        "class", "finally", "long", "strictfp", "volatile",
+        "const", "float", "native", "super", "while",
+        "true", "false", "null",
+    )
+
     override fun generate(element: Element): String = when (element) {
         is File -> emitFile(element)
         else -> emitFile(File(Name.of(""), listOf(element)))
@@ -637,4 +651,4 @@ object JavaGenerator : Generator {
     )
 }
 
-private fun String.sanitize(): String = if (JavaKeywords.reservedKeywords.contains(this)) "_$this" else this
+private fun String.sanitize(): String = if (JavaGenerator.reservedKeywords.contains(this)) "_$this" else this
