+++ src/compiler/ir/src/commonMain/kotlin/community/flock/wirespec/ir/generator/RustGenerator.kt
@@ -58,6 +58,20 @@ import community.flock.wirespec.ir.core.fieldList
 import community.flock.wirespec.ir.core.Function as AstFunction
 
 object RustGenerator : Generator {
+    override val reservedKeywords = setOf(
+        "as", "break", "const", "continue", "crate",
+        "else", "enum", "extern", "false", "fn",
+        "for", "if", "impl", "in", "let",
+        "loop", "match", "mod", "move", "mut",
+        "pub", "ref", "return", "self", "Self",
+        "static", "struct", "super", "trait", "true",
+        "type", "unsafe", "use", "where", "while",
+        "async", "await", "dyn", "abstract", "become",
+        "box", "do", "final", "macro", "override",
+        "priv", "typeof", "unsized", "virtual", "yield",
+        "try",
+    )
+
     override fun generate(element: Element): String = when (element) {
         is File -> element.emit(0)
         else -> File(Name.of(""), listOf(element)).emit(0)
@@ -619,6 +633,6 @@ object RustGenerator : Generator {
 
 // `self` and `Self` cannot be escaped as raw identifiers (r#self is invalid Rust); the emitter's
 // sanitization renames such fields instead, so the generator leaves them untouched.
-private val escapableKeywords = RustKeywords.reservedKeywords - setOf("self", "Self")
+private val escapableKeywords = RustGenerator.reservedKeywords - setOf("self", "Self")
 
 private fun String.sanitize(): String = if (this in escapableKeywords) "r#$this" else this
