+++ src/compiler/ir/src/commonMain/kotlin/community/flock/wirespec/ir/generator/ScalaGenerator.kt
@@ -59,6 +59,18 @@ import community.flock.wirespec.ir.core.fieldList
 import community.flock.wirespec.ir.core.Function as AstFunction
 
 object ScalaGenerator : Generator {
+    override val reservedKeywords = setOf(
+        "abstract", "case", "class", "def", "do",
+        "else", "extends", "false", "final", "for",
+        "forSome", "if", "implicit", "import", "lazy",
+        "match", "new", "null", "object", "override",
+        "package", "private", "protected", "return", "sealed",
+        "super", "this", "throw", "trait", "true",
+        "try", "type", "val", "var", "while",
+        "with", "yield", "given", "using", "enum",
+        "export", "then",
+    )
+
     private var objectNames: Set<String> = emptySet()
     private var primaryFieldNames: Map<String, Set<String>> = emptyMap()
 
@@ -633,4 +645,4 @@ object ScalaGenerator : Generator {
     private fun TypeDescriptor.emitTypeDescriptor(): String = "scala.reflect.classTag[${type.emitGenerics()}]"
 }
 
-private fun String.sanitize(): String = if (ScalaKeywords.reservedKeywords.contains(this)) "`$this`" else this
+private fun String.sanitize(): String = if (ScalaGenerator.reservedKeywords.contains(this)) "`$this`" else this
